+++ CHANGELOG.markdown
@@ -2,6 +2,7 @@
 
  * Fix #28 — Handle BSON BinData properly in documents and ids.
  * Fix — Handle connection auth errors more gracefully.
+ * Add a sanity check for PHP `date.timezone` settings.
 
 
 ## v2.0.2